% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cptac}
\alias{cptac}
\alias{cptac_se}
\alias{cptac_prot}
\alias{cptac_se_prot}
\alias{cptac_na_se}
\title{CPTAC data}
\format{
An object of class \code{MSnSet} with 4051 rows and 6 columns.
}
\usage{
cptac
}
\description{
The data stems from the 6th study of the Clinical Proteomic
Technology Assessment for Cancer (CPTAC). The authors spiked the
Sigma Universal Protein Standard mixture 1 (UPS1) containing 48
different human proteins in a protein background of 60 ng/microL
Saccharomyces cerevisiae strain BY4741. Two different spike-in
concentrations were used: 6A (0.25 fmol UPS1 proteins/microL) and
6B (0.74 fmol UPS1 proteins/microL). In this subset, we limited
ourselves to the data of LTQ-Orbitrap W at site 56. The data were
searched with MaxQuant version 1.5.2.8, and detailed search
settings were described in Goeminne et al. (2016). Three replicates
peptide quantitation data are available for each concentration.
}
\details{
The data are available as \code{MSnSet} or \code{SummarizedExperiment}
objects.

See the proteomics tutorial from the Bioinformatics Summer School
2019 (https://lgatto.github.io/bioc-ms-prot/bss-lab.html) for
scripts on how these data were processed.
}
\examples{
library("MSnbase")
data(cptac)
cptac

## data processing
library("magrittr")
cptac_prot <- cptac \%>\%
   log(base = 2) \%>\%                    ## log2 transformation
   normalise(method = "quantiles") \%>\%  ## normalisation
   combineFeatures(fcol = "Proteins",   ## peptide aggregation
                   method = "robust",
                   na.rm = TRUE)
cptac_prot

## To load a subset of that protein-level data
data(cptac_prot)
cptac_prot

## Same data as SummarizedExperiment object
library("SummarizedExperiment")
data(cptac_se)
cptac_se
data(cptac_se_prot)
cptac_se_prot
}
\keyword{datasets}
