% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g3}
\alias{g3}
\title{Example for distances}
\format{An object of class \code{matrix} with 3 rows and 5 columns.}
\usage{
g3
}
\description{
This is a small toy example providing expression values for 5 genes
and three samples from Olga Vitek. It is used to compare euclidean
and correlation distances and the effect/importance of scaling.
}
\examples{
data(g3)
g3
matplot(t(g3), type = "b", xlab = "Samples", ylab = "Expression")
}
\keyword{datasets}
