% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{giris}
\alias{giris}
\alias{giris2}
\title{Expression data}
\format{
An object of class \code{data.frame} with 150 rows and 5 columns.
}
\usage{
giris
}
\description{
This data is a copy of the \code{iris} data, reframed for biomedical
course. It illustrates the expression of 4 genes, BRCA1, BRCA2,
TP53 and A1CF, in 150 patients, that have been categorised in 3
catégories, A, B and C.
}
\examples{
head(giris)
pairs(giris, col = giris$GRADE)

pca2 <- prcomp(giris2[, -5], scale = TRUE)
factoextra::fviz_pca_ind(pca2, col.ind = giris2$GRADE)
}
\keyword{datasets}
