\name{hiiragi2013d}
\alias{hiiragi2013}
\alias{hiiragi2013df1}
\alias{hiiragi2013df2}
\docType{data}

\title{
  Microarray expression data from Ohnishi et al. 2014.
}


\description{

  The is a subset of the full Hiiragi 2013 dataset from the
  \code{Hiiragi2013} package. The data describes cell-to-cell expression
  variability followed by signal reinforcement progressively segregates
  early mouse lineages.
  
}

\usage{
data("hiiragi2013df1")
data("hiiragi2013df2")
}



\source{

  The data originally come from the \code{Hiiragi2013} Bioconductor
  package. See \code{inst/script/hiiragi2013.R} to see how they have
  been converted.
  
}

\references{

  Cell-to-cell expression variability followed by signal reinforcement
  progressively segregates early mouse lineages by Y. Ohnishi, W. Huber,
  A. Tsumura, M. Kang, P. Xenopoulos, K. Kurimoto, A. K. Oles,
  M. J. Arauzo-Bravo, M.  Saitou, A.-K. Hadjantonakis and T. Hiiragi;
  Nature Cell Biology (2014) 16(1): 27-37. doi: 10.1038/ncb2881.

}


\examples{
data(hiiragi2013df1)
data(hiiragi2013df2)
}


\keyword{datasets}
