% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kem2_se}
\alias{kem2_se}
\title{SummarizedExperiment data}
\format{
An object of class \code{SummarizedExperiment} with 4774 rows and 16 columns.
}
\usage{
kem2_se
}
\description{
This is a small RNA-Seq data set, build from the data returned by
the \code{rWSBIM1207::kem2.tsv()} function.
}
\examples{
library("SummarizedExperiment")
data(kem2_se)
assay(kem2_se)
colData(kem2_se)
rowData(kem2_se)
}
\keyword{datasets}
