% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedata.R
\name{make_data}
\alias{make_data}
\title{Generate data}
\usage{
make_data(noma)
}
\arguments{
\item{noma}{`character(1)` that can be coerced into a numeric}
}
\value{
A SummarizedExperimet
}
\description{
Generate data
}
\examples{
make_data("123")
}
