% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metab1}
\alias{metab1}
\title{Metabolomics data (from MSMB)}
\format{An object of class \code{matrix} with 399 rows and 12 columns.}
\usage{
metab1
}
\description{
This data comes from the Modern \emph{Statistics for Modern Biology} book and was
originally called \code{mat1} from the \code{mat1xcms.RData} file. It contains
quantitation data for 399 metabolites and 6 knock-out and 6 wild-type
samples.
}
\examples{
data(metab1)

dim(metab1)
metab1[1:10, 1:3]
}
\keyword{datasets}
