% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recapSE1}
\alias{recapSE1}
\alias{recapSE2}
\title{Recap SummarizedExperiment data}
\format{
An object of class \code{SummarizedExperiment} with 1550 rows and 20 columns.
}
\usage{
recapSE1
}
\description{
Several \code{SummarizedExperiment} datasets for the recap exercices in
the conclusion chapter of the course.
}
\examples{
data(recapSE1)
recapSE1

data(recapSE2)
recapSE2
}
\keyword{datasets}
