% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWSBIM1322.R
\name{rWSBIM1322version}
\alias{rWSBIM1322version}
\title{Package version}
\usage{
rWSBIM1322version()
}
\description{
Package version
}
\examples{
## check the package version that is currently installed
rWSBIM1322version()
}
