% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tdata1}
\alias{tdata1}
\title{Example data for t-tests illustration}
\format{An object of class \code{matrix} with 100 rows and 6 columns.}
\usage{
tdata1
}
\description{
See `scripts/randata.R` for how these data were generated.
}
\examples{
data(tdata1)
dim(tdata1)
head(tdata1)
}
\keyword{datasets}
