% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xy}
\alias{xy}
\alias{xy0}
\title{Example data for PCA illustration}
\format{An object of class \code{data.frame} with 20 rows and 2 columns.}
\usage{
xy
}
\description{
See \code{man/xy.R} for how these data were generated. \code{xy} is scaled,
\code{xy0} is the orginal data.
}
\examples{
data(xy0)
xy0

data(xy)
xy
pca <- prcomp(xy)
summary(pca)
plot(pca)
biplot(pca)
}
\keyword{datasets}
